Move the final reusable bits of the build app to the core module.
