+++ render/package.json
@@ -18,8 +18,7 @@
   "homepage": "https://github.com/ably/features",
   "dependencies": {
     "@ably/features-core": "file:../core",
-    "jsonschema": "^1.4.0",
-    "yaml": "^1.10.2"
+    "jsonschema": "^1.4.0"
   },
   "devDependencies": {
     "concurrently": "^7.4.0",
